Implement ability to mark board with X and On
Clicking on a square will mark it with either an "X" or an "O",
alternating between the two. Clicking on a non-empty square will
not do anything, nor will it skip the current player's turn. Each
player is represented by a "Player" object, which tracks whether it is
currently that player's turn or not, and what their marker is ("X" or
"O")